+++ include/samurai/petsc/nonlinear_solver.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "fv/cell_based_scheme_assembly.hpp"
 #include "fv/flux_based_scheme_assembly.hpp"
-#include "fv/scheme_operators_assembly.hpp"
+#include "fv/operator_sum_assembly.hpp"
 #include "utils.hpp"
 #include <petsc.h>
